split headers
